+++ CHANGELOG.md
@@ -5,7 +5,8 @@
   (snapshots not signed - only release versions)
 * Add `javaLib` configuration closure:
     - disableGradleMetadata() - disables gradle metadata publishing
-    - bonArtifactId and bomDescription - corrects bom pom when platform declared in the root project
+    - java-platform plugin related configurations (bom sub-closure):
+        * bom.artifactId and bom.description - corrects bom pom when platform declared in the root project
     - disableJavadocPublish() and disableSourcesPublish() - disable javadoc and sources publish, enabled by default
     - enableSnapshotsSigning() - enables signing for snapshot versions (disabled by default)